+++ popgetter-llm/src/embedding.rs
@@ -1,0 +1,1 @@
+// Placeholder
